+++ wrangler.toml
@@ -1,4 +1,5 @@
 name = "image-edit-app"
+main = "src/worker.ts"
 compatibility_date = "2024-04-05"
 
 [[kv_namespaces]]
